Rename the user scoped config file name to jenkins_user.yaml
