+++ controllers/jenkins/config/jenkinsconfig_controller_test.go
@@ -1,4 +1,4 @@
-package jenkinsconfig
+package config
 
 import (
 	"github.com/stretchr/testify/assert"
